+++ src/main/java/com/riakoader/was/webserver/WebServer.java
@@ -13,6 +13,8 @@ public class WebServer {
 
     private static final int DEFAULT_PORT = 8080;
 
+    private static final FrontHandler frontHandler = FrontHandler.getInstance();
+
     public static void main(String[] args) throws Exception {
         int port = 0;
         if (args == null || args.length == 0) {
@@ -21,8 +23,7 @@ public class WebServer {
             port = Integer.parseInt(args[0]);
         }
 
-        WebServerConfig config = WebServerConfig.getInstance();
-        FrontHandler frontHandler = FrontHandler.getInstance();
+        WebServerConfig.getInstance();
 
         try (ServerSocket listenSocket = new ServerSocket(port)) {
             log.info("Web Application Server started {} port.", port);
